requirements.txt
appdirs==1.4.4
asgiref==3.5.0
astroid==2.5.1
attrs==21.2.0
autopep8==1.5.7
awsebcli==3.20.0
bcrypt==3.2.0
blessed==1.18.1
botocore==1.20.112
cached-property==1.5.2
cachetools==5.0.0
cement==2.8.2
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.3
coverage==5.5
cpplint==1.5.4
crispy==0.6.3
cryptography==3.4.7
cycler==0.11.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==4.0.4
django-bootstrap==0.2.4
django-bootstrap3==15.0.0
django-braces==1.14.0
django-ckeditor==6.0.0
django-cors-headers==3.11.0
django-crispy-forms==1.12.0
django-environ==0.8.1
django-filter==2.4.0
django-grappelli==2.14.3
django-import-export==2.5.0
django-ipware==4.0.2
django-jquery==3.1.0
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-nested-admin==3.3.3
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-registration==0.6.3
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
environ==1.0
et-xmlfile==1.1.0
fabio==0.13.0
Faker==8.8.2
filelock==3.0.12
flake8==3.9.2
fonttools==4.33.3
future==0.16.0
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-core==2.3.0
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
iniconfig==1.1.1
isort==5.7.0
jdcal==1.4.1
jedi==0.18.0
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.4.2
lazy-object-proxy==1.5.2
libcst==0.4.4
Markdown==3.3.4
MarkupPy==1.14
matplotlib==3.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
paramiko==2.7.2
parso==0.8.2
pathspec==0.5.9
pbr==5.6.0
pep517==0.10.0
Pillow==9.1.0
pip-tools==6.2.0
pluggy==0.13.1
proto-plus==1.20.3
protobuf==3.20.0
psycopg2-binary==2.9.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.3.0
pylint==2.7.2
PyNaCl==1.4.0
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.2
python-http-client==3.3.7
python-Levenshtein==0.12.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.1.0
python-monkey-business==1.0.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rope==0.19.0
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
semantic-version==2.8.5
sendgrid==6.9.7
silx==1.0.0
six==1.16.0
sko==0.5.7
snowballstemmer==2.1.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stevedore==3.3.0
tablib==2.0.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
typeguard==2.13.3
typing-inspect==0.7.1
typing_extensions==4.2.0
ujson==4.0.2
urllib3==1.26.9
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
waitress==2.0.0
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yapf==0.31.0
